refactor(provider): split AnyAgent impl into run.rs + spawn.rs (dirge-4y4l stage 8)
Decompose the remaining ~800-line AnyAgent impl in provider/mod.rs by
concern into child impl blocks. Behavior-preserving move; child modules
reach AnyAgent's private fields directly (privacy = defining module +
descendants) so no pub(crate) field bumps or accessors were needed.

- provider::run (189 LOC) — run_print, the headless --print/--loop path.
- provider::spawn (429 LOC) — spawn_runner / spawn_review_runner /
  spawn_curator_runner / spawn_memory_curator_runner /
  spawn_review_runner_with_cache / spawn_filtered_runner_with_cache +
  build_stream_fn / build_stream_fn_with_filter.

Builders/accessors (new, extend_loop_tools, with_* x9, provider_name,
cache, loop_tools, memory_provider) stay in mod.rs. provider/mod.rs
951 -> 360 LOC. mod_tests.rs unchanged (super::* still resolves the
inherent methods). -D warnings clean default + windows-default; 2374
tests green.

Refs dirge-7yhq
